% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_biotekfiles.R
\name{import_biotek2_files}
\alias{import_biotek2_files}
\title{Import and clean Biotek2 files}
\usage{
import_biotek2_files(
  path_to_biotekfolder,
  is_googledrive = FALSE,
  download_directory = NA
)
}
\arguments{
\item{path_to_biotekfolder}{A character string specifying the path to the folder containing the \code{.csv} files.}

\item{is_googledrive}{A logical value indicating whether the folder is on Google Drive. Default is \code{FALSE}.}

\item{download_directory}{A character string specifying the directory to download the files to if \code{is_googledrive} is \code{TRUE}. If not specified, a temporary directory will be used. All files in \code{download_directory} will be replaced.}
}
\value{
A data frame containing the cleaned data from all \code{.csv} files.
}
\description{
This function imports and cleans all \code{.csv} files from a specified folder. It can handle files from a local directory or from a Google Drive folder.
}
\examples{
\dontrun{
# Read and clean files from a local directory
data <- read_all_biotek2_files("path/to/local/folder")

# Read and clean files from a Google Drive folder
data <- read_all_biotek2_files("your-google-drive-folder-id", is_googledrive = TRUE)
}
}
