% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_chl_per_area.R
\name{normalise_chl_per_area}
\alias{normalise_chl_per_area}
\title{Normalise Chlorophyll Concentration for Surface Area of Coral Fragments}
\usage{
normalise_chl_per_area(
  data,
  pl = 1,
  v_ml_sw_added = 3,
  v_ml_sw_pipetted = 1,
  path_to_biotekfolder,
  is_googledrive = FALSE,
  download_directory = NA,
  plot = TRUE
)
}
\arguments{
\item{data}{A data frame containing metadata for the samples.}

\item{pl}{A numeric value representing the path length in cm. Default is 1.}

\item{v_ml_sw_added}{A numeric value representing the volume in milliliters of seawater added. Default is 3 mL.}

\item{v_ml_sw_pipetted}{A numeric value representing the volume in milliliters of seawater pipetted. Default is 1 mL.}

\item{path_to_biotekfolder}{A character string representing the path to the folder containing Biotek files.}

\item{is_googledrive}{A logical value indicating whether the Biotek files are stored on Google Drive. Default is FALSE.}

\item{download_directory}{A character string representing the directory to download files if they are stored on Google Drive. Default is NA. All files in that directory will be replaced.}

\item{plot}{A logical value indicating whether to plot the absorption values for blanks and samples. Default is TRUE.}
}
\value{
A data frame with columns \code{sample_id}. \code{chl_a_per_cm2} and \code{chl_c2_per_cm2}, and \code{chl_tot_per_cm2}, representing the chlorophyll a, c2, and total chlorophyll concentration normalised for surface area.
}
\description{
This function calculates the chlorophyll (a + c2) concentration in a coral fragment, normalised for the surface area of that coral fragment, based on absorption measurements. The formula to convert absorptions to chlorophyll concentration comes from Jeffrey & Humphrey (1975).
}
\examples{
\dontrun{
data <- data.frame(
  sample_id = c("sample1", "sample2"),
  type = c("sample", "sample"),
  w1 = c(0.5, 0.5),
  w2 = c(1, 1),
  area = c(10, 15),
  filename = c("file1", "file2"),
  m1 = c(1, 2),
  m2 = c(3, 4)
)
path_to_biotekfolder <- "path/to/biotek/files"
normalise_chla_per_area(data, path_to_biotekfolder = path_to_biotekfolder)
}

}
\references{
Jeffrey, S. W., & Humphrey, G. F. (1975). New spectrophotometric equations for determining chlorophylls a, b, c1 and c2 in higher plants, algae and natural phytoplankton. Biochemie Und Physiologie Der Pflanzen, 167(2), 191–194. \doi{10.1016/S0015-3796(17)30778-3}
}
