% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_counts_per_area.R
\name{normalise_counts_per_area}
\alias{normalise_counts_per_area}
\title{Normalise Cell Counts for Surface Area of Coral Fragments}
\usage{
normalise_counts_per_area(
  data,
  v_ml_count = 0.01,
  v_ml_sw_added = 3,
  v_ml_sw_pipetted = 1
)
}
\arguments{
\item{data}{A data frame containing the cell counts and other relevant measurements. It should have columns \code{c1} to \code{c6} representing count replicates, \code{v_zoox}, \code{v_sw}, \code{w1}, \code{w2}, and \code{area}.}

\item{v_ml_count}{A numeric value representing the volume in milliliters added into the counting chamber. Default is 0.01 mL (= 10 µL).}

\item{v_ml_sw_added}{A numeric value representing the volume in milliliters of seawater added. Default is 3 mL.}

\item{v_ml_sw_pipetted}{A numeric value representing the volume in milliliters of seawater pipetted. Default is 1 mL.}
}
\value{
A data frame with normalised cell counts per surface area of coral fragments.
}
\description{
This function calculates the original number of cells in the tissue slurry and then normalises these counts for the surface area of a coral fragment.
}
\examples{
\dontrun{
data <- data.frame(
  sample_id = c("sample1", "sample2"),
  c1 = c(110, 160),
  c2 = c(105, 155),
  c3 = c(115, 165),
  c4 = c(120, 170),
  c5 = c(125, 175),
  c6 = c(130, 180),
  v_zoox = c(1, 1),
  v_sw = c(2, 2),
  w1 = c(0.5, 0.5),
  w2 = c(1, 1),
  area = c(10, 15)
)
normalise_counts_per_area(data)
}

}
