% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_biotekfiles.R
\name{read_and_clean_file}
\alias{read_and_clean_file}
\title{Read and Clean a Single Biotek2 File}
\usage{
read_and_clean_file(path_to_file)
}
\arguments{
\item{path_to_file}{A character string specifying the path to the \code{.csv} file.}
}
\value{
A data frame containing the cleaned data from the \code{.csv} file.
}
\description{
This function reads and cleans a single \code{.csv} file and is called by \code{read_all_biotek2_files()}.
}
