% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{run_na_tests}
\alias{run_na_tests}
\title{Run a set of test for NA values in coulmns}
\usage{
run_na_tests(data)
}
\arguments{
\item{data}{A data frame to be checked for NA values.}
}
\description{
This fucntion uses the test_for_na() fucntion to run test on the important columns in the combined metadata sheet.
}
