% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_chl_per_area.R
\name{verify_positions}
\alias{verify_positions}
\title{Verify Positions in Data Frames}
\usage{
verify_positions(data_1, data_2)
}
\arguments{
\item{data_1}{A data frame containing columns \code{filename} and \code{position}.}

\item{data_2}{A data frame containing columns \code{filename} and \code{position}.}
}
\value{
A logical value indicating whether all positions in \code{data_1} occur in \code{data_2}.
}
\description{
This function checks if all positions in \code{data_1} occur in \code{data_2} for each filename. If there are positions in \code{data_1} that do not occur in \code{data_2}, it prints a warning with the filename followed by the positions.
}
\examples{
\dontrun{
data_1 <- data.frame(
  filename = c("file1.txt", "file1.txt", "file2.txt", "file2.txt"),
  position = c(1, 2, 1, 4)
)
data_2 <- data.frame(
  filename = c("file1.txt", "file1.txt", "file2.txt", "file2.txt"),
  position = c(1, 2, 1, 3)
)
verify_positions(data_1, data_2)
}

}
