% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_chla_per_area.R
\name{verify_unique_positions}
\alias{verify_unique_positions}
\title{Verify Unique Positions per Filename}
\usage{
verify_unique_positions(data)
}
\arguments{
\item{data}{A data frame containing columns \code{filename} and \code{position}.}
}
\value{
A logical value indicating whether all positions are unique per filename.
}
\description{
This function verifies that each position only occurs once per filename in a data frame. If a duplicated position for one filename is found, it prints the filename followed by the positions.
}
\examples{
\dontrun{
data <- data.frame(
  filename = c("file1.txt", "file1.txt", "file2.txt", "file2.txt", "file2.txt"),
  position = c(1, 2, 1, 1, 3)
)
verify_unique_positions(data)
}

}
