% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area.R
\name{get_area}
\alias{get_area}
\title{Estimate Surface Area of Coral Fragments}
\usage{
get_area(data, conversion_factor = 34.32)
}
\arguments{
\item{data}{A data frame containing the weight measurements of coral fragments. It should have columns \code{sample_id}, \code{w1_scratched}, and \code{w2_scratched}.}

\item{conversion_factor}{A numeric value representing the conversion factor to estimate the surface area from the weight difference. Default is 34.32.}
}
\value{
A data frame with columns \code{sample_id} and \code{area}, where \code{area} is the estimated surface area of the coral fragments.
}
\description{
This function estimates the surface area of coral fragments using the difference in weight with one and two layers of paraffin wax. The conversion factor is based on the study by Veal et al. (2010).
}
\examples{
\dontrun{
data <- data.frame(
  sample_id = c("sample1", "sample2"),
  w1_scratched = c(10, 15),
  w2_scratched = c(12, 18)
)
get_area(data)
}

}
\references{
Veal, C. J., Holmes, G., Nunez, M., Hoegh-Guldberg, O., & Osborn, J. (2010). A comparative study of methods for surface area and three-dimensional shape measurement of coral skeletons. Limnology and Oceanography: Methods, 8(5), 241–253. \doi{10.4319/lom.2010.8.241}
}
