% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Read and Combine Metadata from Google Sheets}
\usage{
read_metadata(googlesheets_link)
}
\arguments{
\item{googlesheets_link}{A character string representing the link to the Google Sheets document.}
}
\value{
A data frame containing combined metadata from the specified Google Sheets.
}
\description{
This function reads metadata from a Google Sheets link and combines data from multiple sheets into a single data frame.
}
\examples{
\dontrun{
googlesheets_link <- "https://docs.google.com/spreadsheets/d/your_sheet_id"
read_metadata(googlesheets_link)
}

}
